% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Input check}
\usage{
check_input(probs)
}
\arguments{
\item{probs}{Error probabilities.}
}
\value{
The result of \code{stopifnot}.
}
\description{
Internal function to check consistency of the two error propabilities.
}
