% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output.R
\name{check_output}
\alias{check_output}
\title{Output Check}
\usage{
check_output(ci, probs, parameter_range = c(-Inf, Inf))
}
\arguments{
\item{ci}{A confidence interval.}

\item{probs}{Error probabilities.}

\item{parameter_range}{Range of parameter domain.}
}
\value{
A confidence interval.
}
\description{
Internal function to squeeze/expand the resulting confidence interval to the parameter domain given error propabilities of 0 or 1.
}
