% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_boot.R
\name{ci_boot}
\alias{ci_boot}
\title{Apply Bootstrap Confidence Interval}
\usage{
ci_boot(S, boot_type = c("bootstrapT", "percentile", "t", "bca"), probs, ...)
}
\arguments{
\item{S}{The result of \code{resample::bootstrap}.}

\item{boot_type}{Type of bootstrap confidence interval ("bootstrapT", "percentile", "t", or "bca").}

\item{probs}{Error probabilities.}

\item{...}{Further arguments passed to \code{resample::CI.boot_type}.}
}
\value{
A confidence interval.
}
\description{
Internal function to calculate a specific type of bootstrap confidence interval from the result of the bootstrap.
}
