% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_skewness.R
\name{ci_skewness}
\alias{ci_skewness}
\title{Confidence Interval for the Skewness}
\usage{
ci_skewness(
  x,
  probs = c(0.025, 0.975),
  type = "bootstrap",
  boot_type = c("percentile", "t", "bca"),
  R = 10000,
  seed = NULL,
  parameter_range = c(-Inf, Inf),
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{Error probabilites. The default c(0.025, 0.975) gives a symmetric 95\% confidence interval.}

\item{type}{Type of confidence interval. Currently not used as the only type is "bootstrap".}

\item{boot_type}{Type of bootstrap confidence interval ("percentile", "t", or "bca").}

\item{R}{The number of bootstrap resamples.}

\item{seed}{An integer random seed.}

\item{parameter_range}{Range of parameter of interest (relevant for the one-sided case).}

\item{...}{Further arguments passed to \code{resample::CI.boot_type}.}
}
\value{
A list with class \code{htest} containing these components:
\itemize{
\item \code{conf.int}: The confidence interval.
\item \code{estimate}: The parameter estimate.
\item \code{method}: A character string describing the applied method.
\item \code{data.name}: A character string with the name(s) of the data.
}
}
\description{
This function calculates bootstrap confidence intervals for the population skewness using the R package "resample". The default Bootstrap type is "percentile". "bootstrapT" is not available since there is no simple general formula for the standard error of the skewness.
}
\details{
Note that for "percentile" and "bca" bootstrap, modified percentiles for better small-sample accuracy are used. Pass \code{expand = FALSE} to \code{...} in order to suppress this.
}
\examples{
set.seed(1)
x <- rnorm(100)
ci_skewness(x, R = 1000)
}
\references{
Tim Hesterberg (2015). resample: Resampling Functions. R package version 0.4. <CRAN.R-project.org/package=resample>.
}
