% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_var.R
\name{ci_var}
\alias{ci_var}
\title{Confidence Interval for the Population Variance}
\usage{
ci_var(
  x,
  probs = c(0.025, 0.975),
  type = c("chi-squared", "bootstrap"),
  boot_type = c("bootstrapT", "percentile", "t", "bca"),
  R = 10000,
  seed = NULL,
  parameter_range = c(0, Inf),
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{probs}{Error probabilites. The default c(0.025, 0.975) gives a symmetric 95\% confidence interval.}

\item{type}{Type of confidence interval. One of "normal" or "bootstrap".}

\item{boot_type}{Type of bootstrap confidence interval ("bootstrapT", "percentile", "t", or "bca"). Only used for \code{type = "bootstrap"}.}

\item{R}{The number of bootstrap resamples. Only used for \code{type = "bootstrap"}.}

\item{seed}{An integer random seed. Only used for \code{type = "bootstrap"}.}

\item{parameter_range}{Range of parameter of interest (relevant for the one-sided case).}

\item{...}{Further arguments passed to \code{resample::CI.boot_type}.}
}
\value{
A list with class \code{htest} containing these components:
\itemize{
\item \code{conf.int}: The confidence interval.
\item \code{estimate}: The parameter estimate.
\item \code{method}: A character string describing the applied method.
\item \code{data.name}: A character string with the name(s) of the data.
}
}
\description{
This function calculates confidence intervals for the population variance. By default, confidence intervals based on the chi-squared distribution are used. Alternatively, bootstrap confidence intervals are supported by the package "resample". The default Bootstrap type is "BootstrapT". It uses the standard error for the variance given in Wilks (see references).
}
\details{
Note that for "percentile" and "bca" bootstrap, modified percentiles for better small-sample accuracy are used. Pass \code{expand = FALSE} to \code{...} in order to suppress this.
}
\examples{
x <- 1:100
ci_var(x)
ci_var(x, type = "bootstrap", R = 1000)
ci_var(x, type = "bootstrap", boot_type = "bca", R = 1000)
}
\references{
\enumerate{
\item S.S. Wilks, Mathematical Statistics, Wiley & Sons, page 199, (8.2.9).
\item Tim Hesterberg (2015). resample: Resampling Functions. R package version 0.4. <CRAN.R-project.org/package=resample>.
}
}
