% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output.R
\name{prepare_output}
\alias{prepare_output}
\title{Preparation of Output}
\usage{
prepare_output(ci, estimate, probs, type, boot_type, data_name, estimate_name)
}
\arguments{
\item{ci}{A numeric vector of length two representing a confidence interval.}

\item{estimate}{A numeric vector of length one representing the estimate.}

\item{probs}{Error probabilites.}

\item{type}{Type of confidence interval.}

\item{boot_type}{Type of bootstrap confidence interval.}

\item{data_name}{Name of the input data.}

\item{estimate_name}{Name of the estimate.}
}
\value{
A list with class \code{htest} containing these components:
\itemize{
\item \code{conf.int}: The confidence interval.
\item \code{estimate}: The parameter estimate.
\item \code{method}: A character string describing the applied method.
\item \code{data.name}: A character string with the name(s) of the data.
}
}
\description{
Internal function to organize output of confidence interval functions.
}
