% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderr_var.R
\name{stderr_var}
\alias{stderr_var}
\title{Standard error of variance}
\usage{
stderr_var(z)
}
\arguments{
\item{z}{A numeric vector.}
}
\value{
A numeric vector of length 1.
}
\description{
Internal function to calculate the standard error for the variance. It is used in \code{resample::bootstrap} and needs to be exported for this purpose.
}
\examples{
stderr_var(1:100)
}
